refactor: extract neural search pipeline into @repo/console-search
Move 9 neural search files from apps/console/src/lib/neural/ into
packages/console-search/ so the eval pipeline can import searchLogic()
in-process instead of going through HTTP.

- Create @repo/console-search package with proper workspace conventions
- Move neural files (four-path-search, entity-search, cluster-search,
  actor-search, llm-filter, url-builder, url-resolver, id-resolver,
  temporal-state) via git mv
- Extract searchLogic() with injectable onActivity callback
- Update v1 files (search, contents, findsimilar, index) to import
  from @repo/console-search
- Add @repo/console-search dependency to apps/console
