+++ packages/console-search/package.json
@@ -1,0 +1,46 @@
+{
+  "name": "@repo/console-search",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./src/index.ts"
+    }
+  },
+  "license": "MIT",
+  "scripts": {
+    "build": "tsup",
+    "clean": "git clean -xdf .cache .turbo dist node_modules",
+    "dev": "tsup --watch",
+    "format": "prettier --check . --ignore-path ../../.gitignore",
+    "lint": "eslint",
+    "typecheck": "tsc --noEmit"
+  },
+  "dependencies": {
+    "@ai-sdk/gateway": "catalog:",
+    "@db/console": "workspace:*",
+    "@repo/console-embed": "workspace:*",
+    "@repo/console-pinecone": "workspace:*",
+    "@repo/console-rerank": "workspace:*",
+    "@repo/console-types": "workspace:*",
+    "@repo/console-validation": "workspace:*",
+    "@repo/console-workspace-cache": "workspace:*",
+    "@vendor/observability": "workspace:*",
+    "ai": "catalog:",
+    "drizzle-orm": "catalog:",
+    "zod": "catalog:zod3"
+  },
+  "devDependencies": {
+    "@repo/eslint-config": "workspace:*",
+    "@repo/prettier-config": "workspace:*",
+    "@repo/typescript-config": "workspace:*",
+    "@types/node": "catalog:",
+    "eslint": "catalog:",
+    "prettier": "catalog:",
+    "tsup": "^8.5.0",
+    "typescript": "catalog:"
+  },
+  "prettier": "@repo/prettier-config"
+}
