+++ packages/console-search/src/neural/four-path-search.ts
