+++ packages/console-search/src/neural/temporal-state.ts
