+++ packages/console-search/src/types.ts
@@ -1,0 +1,7 @@
+/** Auth context available to all v1 logic functions */
+export interface V1AuthContext {
+  workspaceId: string;
+  userId: string;
+  authType: "api-key" | "session";
+  apiKeyId?: string;
+}
